Split demo course content dropdown
